# openapi.yaml
openapi: 3.1.0
info:
  title: KAKAoke API
  description: |
    Backend API for the KAKAoke karaoke web application.
    Serves UltraStar song metadata, synchronized lyrics, and media assets
    from a local song library.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
  /api/auth/register:
    post:
      operationId: register
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registration successful. Returns JWT token and user info.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: |
            Invalid registration. Username must be 3-50 characters,
            password must be 6+ characters, or username already taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Log in with username and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful. Returns JWT token and user info.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid username or password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/me:
    get:
      operationId: getCurrentUser
      tags: [Auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user info.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated or token expired.
  /api/songs:
    get:
      operationId: getSongs
      tags: [Songs]
      summary: List songs in the user's library
      description: |
        Returns a paginated list of songs from the current user's library.
        Supports full-text search across title and artist, filtering by genre,
        language, and year, and sorting.
      parameters:
        - name: search
          in: query
          description: |
            Free-text search query. Matches against song title and artist name
            (case-insensitive, partial match).
          schema:
            type: string
          example: "Bohemian"
        - name: genre
          in: query
          description: Filter by genre (case-insensitive exact match).
          schema:
            type: string
          example: "Rock"
        - name: edition
          in: query
          description: Filter by edition (case-insensitive exact match).
          schema:
            type: string
          example: "Charts"
        - name: language
          in: query
          description: Filter by language (case-insensitive exact match).
          schema:
            type: string
          example: "English"
        - name: year
          in: query
          description: Filter by release year.
          schema:
            type: integer
          example: 1975
        - name: favorite
          in: query
          description: When true, only return songs the user has marked as favourite.
          schema:
            type: boolean
        - name: sortBy
          in: query
          description: Field to sort results by.
          schema:
            type: string
            enum: [title, artist, year]
            default: title
        - name: sortOrder
          in: query
          description: Sort direction.
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page
          in: query
          description: Zero-based page index.
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Number of songs per page.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Paginated list of songs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongPage"
  /api/songs/{songId}:
    get:
      operationId: getSongById
      tags: [Songs]
      summary: Get a single song by ID
      parameters:
        - $ref: "#/components/parameters/songId"
      responses:
        "200":
          description: Song metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateSong
      tags: [Songs]
      summary: Update song metadata
      description: |
        Updates the metadata of an existing song. Only provided fields are
        updated — omitted fields retain their current values. To clear an
        optional field, send an empty string (for text) or an empty array
        (for tags).
      parameters:
        - $ref: "#/components/parameters/songId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongUpdate"
      responses:
        "200":
          description: Updated song metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeSongFromLibrary
      tags: [Songs]
      summary: Remove a song from the user's library
      description: |
        Removes the song from the current user's library. The song remains in
        the global pool and is still available to other users who have it. This
        does not delete the actual song files.
      parameters:
        - $ref: "#/components/parameters/songId"
      responses:
        "204":
          description: Song removed from library.
        "404":
          $ref: "#/components/responses/NotFound"
  /api/songs/{songId}/lyrics:
    get:
      operationId: getLyricTimeline
      tags: [Lyrics]
      summary: Get the synchronized lyric timeline for a song
      description: |
        Returns the full lyric timeline with millisecond-accurate timing
        derived from the UltraStar note data. Each line contains tokens
        (syllables) with start time, duration, pitch, and note type.
      parameters:
        - $ref: "#/components/parameters/songId"
      responses:
        "200":
          description: Lyric timeline with timed tokens grouped into lines.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LyricTimeline"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/songs/{songId}/played:
    post:
      operationId: markSongPlayed
      tags: [Songs]
      summary: Mark a song as played
      description: |
        Sets the `played` flag to `true` for the current user's association
        with this song. This removes the "New" badge from the library grid.
        Per-user — other users' played status is unaffected. Idempotent.
      parameters:
        - $ref: "#/components/parameters/songId"
      responses:
        "204":
          description: Song marked as played.
        "404":
          $ref: "#/components/responses/NotFound"
  /api/songs/{songId}/favorite:
    post:
      operationId: favoriteSong
      tags: [Songs]
      summary: Add a song to favorites
      description: |
        Marks the song as a favorite for the current user. Per-user — does not
        affect other users. Idempotent.
      parameters:
        - $ref: "#/components/parameters/songId"
      responses:
        "204":
          description: Song added to favorites.
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: unfavoriteSong
      tags: [Songs]
      summary: Remove a song from favorites
      description: |
        Removes the song from the current user's favorites. Per-user — does not
        affect other users. Idempotent.
      parameters:
        - $ref: "#/components/parameters/songId"
      responses:
        "204":
          description: Song removed from favorites.
        "404":
          $ref: "#/components/responses/NotFound"
  /api/songs/import:
    post:
      operationId: importSong
      tags: [Import]
      summary: Import a song from a ZIP archive
      description: |
        Accepts a ZIP file containing an UltraStar `.txt` file and its
        associated assets (audio, cover, background). The server extracts
        the archive, parses the UltraStar file, generates a cover thumbnail,
        uploads the `.txt` file synchronously, and saves the song to the
        database with `status: processing`. Large files (audio, cover,
        background) are uploaded to storage asynchronously in the background.

        The response returns immediately with the song metadata and thumbnail
        URL. The `status` field will be `processing` until all files finish
        uploading, then it transitions to `ready`. If the upload fails, the
        status becomes `broken`.

        **Deduplication:** The server computes a SHA-256 hash of the `.txt`
        file content. If an existing song has the same hash and is ready,
        it is simply added to the current user's library (no re-upload).
        If re-importing a broken entry, the song is re-uploaded and repaired.

        The ZIP must contain exactly one `.txt` file at any level.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: ZIP archive containing the UltraStar song and assets.
      responses:
        "201":
          description: Song successfully imported.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "400":
          description: |
            Invalid archive. Possible reasons: not a valid ZIP, no `.txt` file
            found, multiple `.txt` files found, or UltraStar parse error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    songId:
      name: songId
      in: path
      required: true
      description: Unique identifier of the song.
      schema:
        type: string
  responses:
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/api/auth/login` or `/api/auth/register`.
        Include in the `Authorization` header as `Bearer <token>`.
  schemas:
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Unique username for the new account.
          example: "singer42"
        password:
          type: string
          minLength: 6
          description: Password (minimum 6 characters).
          example: "mypassword"
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: Username.
          example: "singer42"
        password:
          type: string
          description: Password.
          example: "mypassword"
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT authentication token.
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, admin]
      properties:
        id:
          type: integer
          format: int64
          description: Unique user ID.
          example: 1
        username:
          type: string
          description: Username.
          example: "singer42"
        admin:
          type: boolean
          description: Whether the user has admin privileges.
          default: false
    Song:
      type: object
      required: [id, title, artist, bpm, gap]
      properties:
        id:
          type: string
          description: Unique identifier for the song.
          example: "queen-bohemian-rhapsody"
        title:
          type: string
          description: Song title (from `#TITLE` header).
          example: "Bohemian Rhapsody"
        artist:
          type: string
          description: Artist name (from `#ARTIST` header).
          example: "Queen"
        genre:
          type: string
          nullable: true
          description: Genre (from `#GENRE` header), null if not specified.
          example: "Rock"
        year:
          type: integer
          nullable: true
          description: Release year (from `#YEAR` header), null if not specified.
          example: 1975
        language:
          type: string
          nullable: true
          description: Language (from `#LANGUAGE` header), null if not specified.
          example: "English"
        edition:
          type: string
          nullable: true
          description: Edition tag (from `#EDITION` header), null if not specified.
          example: "Charts"
        creator:
          type: string
          nullable: true
          description: Who created the UltraStar file (from `#CREATOR` header).
        bpm:
          type: number
          format: double
          description: |
            Beats per minute as stored in the file (before the implicit 4x
            multiplication). Used for beat-to-millisecond conversion.
          example: 71.5
        gap:
          type: number
          format: double
          description: |
            Offset in milliseconds from audio start to beat 0
            (from `#GAP` header). Defaults to 0 if absent.
          example: 5320.0
        duet:
          type: boolean
          description: Whether this is a multi-voice (duet) song.
          default: false
        voiceNames:
          type: array
          nullable: true
          description: |
            Names of the voices for duet songs (from `#P1`–`#P9` headers).
            Null for single-voice songs.
          items:
            type: string
          example: ["Freddie", "Roger"]
        previewStart:
          type: number
          format: double
          nullable: true
          description: |
            Offset in seconds where the song preview should start
            (from `#PREVIEWSTART` header).
        coverUrl:
          type: string
          format: uri
          nullable: true
          description: |
            URL to retrieve the full-size cover image. Null if no cover exists
            or if the song is still processing (assets not yet uploaded).
          example: "/api/songs/queen-bohemian-rhapsody/cover"
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
          description: |
            URL to retrieve a small (~200px) JPEG thumbnail of the cover.
            Available immediately after import (stored in DB). Preferred for
            library grid views to save bandwidth.
          example: "/api/songs/queen-bohemian-rhapsody/thumbnail"
        backgroundUrl:
          type: string
          format: uri
          nullable: true
          description: URL to retrieve the background image, or null if none exists.
          example: "/api/songs/queen-bohemian-rhapsody/background"
        audioUrl:
          type: string
          format: uri
          nullable: true
          description: URL to stream the audio file, or null if no audio exists.
          example: "/api/songs/queen-bohemian-rhapsody/audio"
        videoUrl:
          type: string
          format: uri
          nullable: true
          description: |
            URL to stream the video file (from `#VIDEO` header), or null if
            no video exists. Supports HTTP Range requests for streaming.
          example: "/api/songs/queen-bohemian-rhapsody/video"
        videoGap:
          type: number
          format: double
          nullable: true
          description: |
            Offset in seconds between the start of the video and the start of
            the audio (from `#VIDEOGAP` header). Positive values mean the video
            starts later than the audio. Null if not specified.
          example: 0.0
        tags:
          type: array
          nullable: true
          description: User-defined tags for categorization.
          items:
            type: string
          example: ["classic", "favorite", "party"]
        status:
          type: string
          enum: [ready, processing, broken]
          description: |
            Current state of the song:
            - `ready`: All files uploaded and available for playback.
            - `processing`: Song was just imported and large files (audio, cover,
              background) are still uploading to storage in the background.
              Metadata and thumbnail are available; asset URLs are null.
            - `broken`: Essential files are missing (e.g., .txt file lost).
          default: ready
          example: "ready"
        played:
          type: boolean
          description: |
            Whether this song has been played at least once by the current user.
            Per-user — each user has their own played status. Songs that have
            never been played show a "New" badge in the library grid.
          default: false
        favorite:
          type: boolean
          description: |
            Whether the current user has favorited this song. Per-user — each
            user has their own favorites.
          default: false
    SongUpdate:
      type: object
      description: |
        Partial song metadata update. Only provided fields are updated.
        Send an empty string to clear a text field, or an empty array to
        clear tags.
      properties:
        title:
          type: string
          description: Updated song title.
        artist:
          type: string
          description: Updated artist name.
        genre:
          type: string
          description: Updated genre. Empty string clears it.
        edition:
          type: string
          description: Updated edition. Empty string clears it.
        creator:
          type: string
          description: Updated creator name. Empty string clears it.
        year:
          type: integer
          nullable: true
          description: Updated release year. Null clears it.
        language:
          type: string
          description: Updated language. Empty string clears it.
        tags:
          type: array
          description: Updated tags. Empty array clears them.
          items:
            type: string
    SongPage:
      type: object
      required: [content, totalElements, totalPages, page, size]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Song"
          description: Songs on the current page.
        totalElements:
          type: integer
          description: Total number of songs matching the query.
          example: 142
        totalPages:
          type: integer
          description: Total number of pages.
          example: 3
        page:
          type: integer
          description: Current zero-based page index.
          example: 0
        size:
          type: integer
          description: Page size.
          example: 50
    LyricTimeline:
      type: object
      required: [songId, lines]
      properties:
        songId:
          type: string
          description: ID of the song this timeline belongs to.
          example: "queen-bohemian-rhapsody"
        lines:
          type: array
          description: Ordered list of lyric lines (phrases).
          items:
            $ref: "#/components/schemas/LyricLine"
    LyricLine:
      type: object
      required: [tokens, startMs]
      properties:
        startMs:
          type: number
          format: double
          description: Start time of this line in milliseconds from audio start.
          example: 5320.0
        tokens:
          type: array
          description: |
            Ordered syllable tokens within this line. Concatenating all
            token texts reproduces the full lyric text for the line.
          items:
            $ref: "#/components/schemas/LyricToken"
    LyricToken:
      type: object
      required: [text, startMs, durationMs, pitch, noteType]
      properties:
        text:
          type: string
          description: |
            Syllable text, including any leading/trailing spaces that separate
            it from adjacent syllables.
          example: "Is "
        startMs:
          type: number
          format: double
          description: Start time of this token in milliseconds from audio start.
          example: 5320.0
        durationMs:
          type: number
          format: double
          description: Duration of this token in milliseconds.
          example: 420.0
        pitch:
          type: integer
          description: |
            MIDI-relative pitch as half-steps from C4 (middle C).
            Ignored for freestyle and rap notes.
          example: 5
        noteType:
          $ref: "#/components/schemas/NoteType"
    NoteType:
      type: string
      description: |
        UltraStar note type indicating scoring behavior:
        - `:` Normal note — standard pitch scoring
        - `*` Golden note — awards bonus points
        - `F` Freestyle — not scored
        - `R` Rap — scored on timing only, pitch ignored
        - `G` Golden rap — rap with bonus points
      enum: [":", "*", "F", "R", "G"]
...
    Error:
      type: object
      required: [status, message]
      properties:
        status:
          type: integer
          description: HTTP status code.
          example: 404
        message:
          type: string
          description: Human-readable error description.
          example: "Song not found"
        detail:
          type: string
          nullable: true
          description: Additional detail about the error, if available.